wc: improve byte count performance
Previously, wc would always read in the file before acting upon the
arguments that were passed into it.

This didn't make sense for the -c flag, as the number of bytes is stored
in the metadata for the file, meaning it doesn't have to be read in at
all.

This commit changse the ordering of events so that the reading of the
file doesn't happen until it absolutely needs to happen; in the branches
where wc needs to return line-related information, and in the default
case.

Therefore, wc can short-circuit the reading of the file when the user
requests the byte-count, and returns the count instantly.