wc: add support for any-ordered arguments
Previously, the implementation for parsing arguments relied on the
positioning, but this meant that users were forced to specify the
arguments in a particular manner.

Now, wc uses some very rudimentary checking to find the file name and
the argument that was passed in, no matter the order of the two.

NOTE: wc can still only handle one argument at a time, and that's a
feature that will need to be added in a future patch.