wc: correct line count for empty files
The bash implementation of wc doesn't count lines that only contain the
NULL terminator as valid lines, and we weren't accounting for that here.

Luckily, as far as I can tell, we only have to do this for files that
_only_ have the NULL character in them (so only have one line).

I'm sure this is going to come back and bite us in the future, but this
at least lines up the output with what wc spits out for the small test
cases that I was trying with.